Create wp-admin-notes
Two functions for WordPress that enable a specified post to be pinned at the top of the admin posts list and hidden from the front end when published. This allows for a post to be used for notes about the site. These functions should be placed in the functions.php file in the theme or theme child folder.